langchain/24-chatbot-service/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.135.3",
    "langchain>=1.2.15",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "langchain-postgres>=0.0.17",
    "langgraph>=1.1.6",
    "langgraph-checkpoint-postgres>=3.0.5",
    "psycopg[binary]>=3.3.3",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "sqlalchemy>=2.0.49",

]
